requirements.txt
black==22.1.0
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-decouple==3.6
django-cors-headers==3.11.0
gunicorn==20.1.0